refactor(core): 收口自动压缩完成处理
